Fix method name. Rename to data.py
