+++ atm/__init__.py
@@ -15,7 +15,7 @@ __email__ = 'dailabmit@gmail.com'
 __version__ = '0.1.3-dev'
 
 # this defines which modules will be imported by "from atm import *"
-__all__ = ['config', 'classifier', 'constants', 'database', 'enter_data',
+__all__ = ['config', 'classifier', 'constants', 'data', 'database',
            'method', 'metrics', 'models', 'utilities', 'worker']
 
 # by default, nothing should be logged
